Add generic graphic output over rdp
This option enables an rdp server for one primary display.
It is intended to be used e.g. with the aws-ec2-x86-64 or
aws-ec2-arm64 machines in the first place but can be useful
for debugging purposes for local setup as well.

Bug-AGL: SPEC-5093

Change-Id: I2659628d389080ba382da3c7e5ec1286c2ed1286
Signed-off-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
(cherry picked from commit 405183531725ec7ed27985bc62e2804658b4e3d0)